requirements.txt
psutil
pynvml
distro
python-dotenv
python-telegram-bot[job-queue]
aiohttp
openai
pydub
feedparser
openai-whisper
unalix